backend/requirements.txt
google-genai>=1.0.0
sentence-transformers>=2.2.2
qdrant-client>=1.7.0
numpy>=1.24.0
tqdm>=4.65.0
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
pydantic>=2.5.0
python-dotenv>=1.0.0
rich>=13.0.0
rank-bm25>=0.2.2
psycopg2-binary>=2.9.9
bcrypt>=4.0.0
PyJWT>=2.8.0
PyPDF2>=3.0.0
python-docx>=1.0.0
python-multipart>=0.0.6
